requirements.txt
langchain
langchain-community
langchain-text-splitters
pypdf
langchain-ollama
python-dotenv
pinecone
langchain-groq